Update API base URLs and refactor AllUsers component
